Backend/requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
Flask==2.2.3
Flask-MySQL==1.5.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyOpenSSL==23.0.0
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
